Refactor(api): Implement Service Layer pattern to separate business logic from Controller
